+++ rnote-engine/src/meson.build
@@ -3,10 +3,11 @@ rnote_engine_sources = files(
     'document/background.rs',
     'document/format.rs',
     'document/mod.rs',
-    'engine/visual_debug.rs',
+    'engine/mod.rs',
     'engine/export.rs',
     'engine/import.rs',
-    'engine/mod.rs',
+    'engine/visual_debug.rs',
+    'engine/rendering.rs',
     'pens/mod.rs',
     'pens/penholder.rs',
     'pens/brush.rs',
